fix(inventory): split consumption pure core from DB orchestration (test resolve)
The `test` job failed: consumption.test.ts imports consumption.ts, which imported
@/lib/prisma at module top — vitest's resolver doesn't alias "@/lib/*", so the
import threw before any test ran. Made consumption.ts pure (aggregateConsumption /
channelWeight / mytDayWindow / types — no DB imports) and moved the prisma-backed
postOutletConsumption into consumption-post.ts. Cron route imports from both.
